+++ test/unit/fuzz-tools.test.js
@@ -371,6 +371,7 @@ import * as hardyAdenomaV391 from '../../lib/hardy-adenoma-v391.js';
 import * as hillFlapValveV392 from '../../lib/hill-flap-valve-v392.js';
 import * as laurenGastricV393 from '../../lib/lauren-gastric-v393.js';
 import * as borrmannGastricV394 from '../../lib/borrmann-gastric-v394.js';
+import * as parksFistulaV395 from '../../lib/parks-fistula-v395.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -687,6 +688,7 @@ const MODULES = {
   'hill-flap-valve-v392.js': hillFlapValveV392,
   'lauren-gastric-v393.js': laurenGastricV393,
   'borrmann-gastric-v394.js': borrmannGastricV394,
+  'parks-fistula-v395.js': parksFistulaV395,
   'tb-testing.js': tbTesting,
 };